+++ internal/testenv/docdrift_packages_test.go
@@ -39,6 +39,7 @@ var canonicalDriftTests = []string{
 	"TestThePagesHeadersRestateEverySecurityHeaderTheNginxConfigDeclares",
 	"TestDocVerifyCommandsFailLoudlyAndNameTheirURL",
 	"TestDocKeyUnreadableCountReadsTheWireShape",
+	"TestDocPrometheusExampleScrapesOverTLS",
 }
 
 func TestTheDocsOnlyPathStillRunsTheDriftGuards(t *testing.T) {